refactoring

+++ backend/utils/uploadImage.js
@@ -12,47 +12,42 @@ export const uploadImage = async (
   folderName,
   originalFileName
 ) => {
-  if (originalFileName !== 'avatar-defaultProfil.jpg') {
-    const filePath = path.resolve(
-      `client/public/uploads/${folderName}/${originalFileName}`
+  if (
+    file.detectedFileExtension === '.jpg' ||
+    file.detectedFileExtension === '.png' ||
+    file.detectedFileExtension === '.jpeg' ||
+    file.detectedFileExtension === '.gif'
+  ) {
+    console.log(
+      `The received file is an image. Extension ${file.detectedFileExtension} accepted`
     );
-    fs.unlink(filePath, (err) => {
-      if (err) {
-        console.log('failed to delete local image:' + err);
-      } else {
-        console.info(`Successfully removed file with the path of ${filePath}`);
-      }
-    });
+
+    if (originalFileName !== 'avatar-defaultProfil.jpg') {
+      const filePath = path.resolve(
+        `client/public/uploads/${folderName}/${originalFileName}`
+      );
+      fs.unlink(filePath, (err) => {
+        if (err) {
+          console.log('Failed to delete local image:' + err);
+        } else {
+          console.info(
+            `Successfully removed file with the path of ${filePath}`
+          );
+        }
+      });
+    }
+
+    const fileName = startFileName + '-' + uuidv4();
+
+    await pipeline(
+      file.stream,
+      fs.createWriteStream(
+        `${__dirname}/../client/public/uploads/${folderName}/${fileName}`
+      )
+    );
+
+    return fileName;
+  } else {
+    return false;
   }
-  // try {
-  //   if (
-  //     !file.detectedFileExtension  == '.jpg' ||
-  //     !file.detectedFileExtension  == '.png' ||
-  //     !file.detectedFileExtension  == '.jpeg'
-  //   )
-  //     throw Error('invalid file');
-  // } catch (error) {
-  //   return { message: 'We failed to update your image for some reason...' };
-  // }
-  // if (
-  //   file.detectedFileExtension == '.jpg' ||
-  //   file.detectedFileExtension == '.png' ||
-  //   file.detectedFileExtension == '.jpeg' ||
-  //   file.detectedFileExtension == '.gif'
-  // ) {
-  //   console.log('OK !!!!!');
-  // } else {
-  //   console.log('PROBLEME !!!!!');
-  // }
-
-  const fileName = startFileName + '-' + uuidv4();
-
-  await pipeline(
-    file.stream,
-    fs.createWriteStream(
-      `${__dirname}/../client/public/uploads/${folderName}/${fileName}`
-    )
-  );
-
-  return fileName;
 };
